re #2428, moved livereload operation into a developer command and out of packages.py
